+++ crates/astria-sequencer/src/accounts/mod.rs
@@ -1,5 +1,5 @@
+pub(crate) mod action;
 pub(crate) mod component;
 pub(crate) mod query;
 pub(crate) mod state_ext;
-pub(crate) mod transaction;
 pub(crate) mod types;
