+++ crates/astria-sequencer/src/secondary/mod.rs
@@ -1,1 +1,1 @@
-pub(crate) mod transaction;
+pub(crate) mod action;
